+++ src/notation/kif/kif.ts
@@ -3,7 +3,7 @@ import { Board } from '../../board.js';
 import { findHandicap, isHandicap } from '../../handicaps.js';
 import { Hand, Hands } from '../../hands.js';
 import { initialSfen, makeSfen, parseSfen } from '../../sfen.js';
-import { Color, Move, Rules, Square, isDrop, isNormal } from '../../types.js';
+import { Color, MoveOrDrop, Rules, Square, isDrop, isMove } from '../../types.js';
 import { defined, parseCoordinates } from '../../util.js';
 import { Position } from '../../variant/position.js';
 import { allRoles, dimensions, handRoles, promote } from '../../variant/util.js';
@@ -262,8 +262,8 @@ export function normalizedKifLines(kif: string): string[] {
 
 export const chushogiKifMoveRegex =
   /((?:(?:[１２３４５６７８９]{1,2}|\d\d?)(?:十?[一二三四五六七八九十]))|仝|同)(\S{1,2})((?:（居食い）)|不成|成)?\s?[（|\(|←]*((?:[１２３４５６７８９]{1,2}|\d\d?)(?:十?[一二三四五六七八九十]))[）|\)]/;
-function parseChushogiMove(kifMove: string, lastDest: Square | undefined = undefined): Move | undefined {
-  const match = kifMove.match(chushogiKifMoveRegex);
+function parseChushogiMove(kifMd: string, lastDest: Square | undefined = undefined): MoveOrDrop | undefined {
+  const match = kifMd.match(chushogiKifMoveRegex);
   if (match) {
     const dest = parseJapaneseSquare(match[1]) ?? lastDest;
     if (!defined(dest)) return;
@@ -281,10 +281,10 @@ export const kifMoveRegex =
   /((?:[１２３４５６７８９][一二三四五六七八九]|同\s?))(玉|飛|龍|角|馬|金|銀|成銀|桂|成桂|香|成香|歩|と)(不成|成)?\(([1-9][1-9])\)/;
 export const kifDropRegex = /((?:[１２３４５６７８９][一二三四五六七八九]|同\s?))(飛|角|金|銀|桂|香|歩)打/;
 
-// Parsing kif moves
-export function parseKifMove(kifMove: string, lastDest: Square | undefined = undefined): Move | undefined {
-  // Normal move
-  const match = kifMove.match(kifMoveRegex);
+// Parsing kif moves/drops
+export function parseKifMoveOrDrop(kifMd: string, lastDest: Square | undefined = undefined): MoveOrDrop | undefined {
+  // Move
+  const match = kifMd.match(kifMoveRegex);
   if (match) {
     const dest = parseJapaneseSquare(match[1]) ?? lastDest;
     if (!defined(dest)) return;
@@ -296,8 +296,8 @@ export function parseKifMove(kifMove: string, lastDest: Square | undefined = und
     };
   } else {
     // Drop
-    const match = kifMove.match(kifDropRegex);
-    if (!match || !match[1]) return parseChushogiMove(kifMove, lastDest);
+    const match = kifMd.match(kifDropRegex);
+    if (!match || !match[1]) return parseChushogiMove(kifMd, lastDest);
 
     return {
       role: kanjiToRole(match[2])[0]!,
@@ -306,60 +306,60 @@ export function parseKifMove(kifMove: string, lastDest: Square | undefined = und
   }
 }
 
-function isLionDouble(kifMove: string | undefined): boolean {
-  const m = defined(kifMove) ? (kifMove || '').split('*')[0].trim() : '';
+function isLionDouble(kifMd: string | undefined): boolean {
+  const m = defined(kifMd) ? (kifMd || '').split('*')[0].trim() : '';
   return m.includes('一歩目') || m.includes('二歩目');
 }
 
-export function parseKifMoves(kifMoves: string[], lastDest: Square | undefined = undefined): Move[] {
-  const moves: Move[] = [];
-  for (let i = 0; i < kifMoves.length; i++) {
-    const m = kifMoves[i];
-    let move: Move | undefined;
-    if (isLionDouble(m) && isLionDouble(kifMoves[i + 1])) {
+export function parseKifMovesOrDrops(kifMds: string[], lastDest: Square | undefined = undefined): MoveOrDrop[] {
+  const mds: MoveOrDrop[] = [];
+  for (let i = 0; i < kifMds.length; i++) {
+    const m = kifMds[i];
+    let md: MoveOrDrop | undefined;
+    if (isLionDouble(m) && isLionDouble(kifMds[i + 1])) {
       const firstMove = parseChushogiMove(m),
-        secondMove = parseChushogiMove(kifMoves[++i]);
-      if (firstMove && secondMove && isNormal(firstMove) && isNormal(secondMove)) {
-        move = { from: firstMove.from, to: secondMove.to, midStep: firstMove.to, promotion: false };
+        secondMove = parseChushogiMove(kifMds[++i]);
+      if (firstMove && secondMove && isMove(firstMove) && isMove(secondMove)) {
+        md = { from: firstMove.from, to: secondMove.to, midStep: firstMove.to, promotion: false };
       }
-    } else move = parseKifMove(m, lastDest);
-    if (!move) return moves;
-    lastDest = move.to;
-    moves.push(move);
+    } else md = parseKifMoveOrDrop(m, lastDest);
+    if (!md) return mds;
+    lastDest = md.to;
+    mds.push(md);
   }
-  return moves;
+  return mds;
 }
 
-// Making kif formatted moves
-export function makeKifMove(pos: Position, move: Move, lastDest?: Square): string | undefined {
+// Making kif formatted moves/drops
+export function makeKifMoveOrDrop(pos: Position, md: MoveOrDrop, lastDest?: Square): string | undefined {
   const ms = pos.rules === 'chushogi' ? makeJapaneseSquareHalf : makeJapaneseSquare;
-  if (isDrop(move)) {
-    return ms(move.to) + roleToKanji(move.role) + '打';
+  if (isDrop(md)) {
+    return ms(md.to) + roleToKanji(md.role) + '打';
   } else {
     const sameSquareSymbol = pos.rules === 'chushogi' ? '仝' : '同　',
-      sameDest = (lastDest ?? pos.lastMove?.to) === move.to,
-      moveDestStr = sameDest ? sameSquareSymbol : ms(move.to),
-      promStr = move.promotion ? '成' : '',
-      role = pos.board.getRole(move.from);
+      sameDest = (lastDest ?? pos.lastMoveOrDrop?.to) === md.to,
+      moveDestStr = sameDest ? sameSquareSymbol : ms(md.to),
+      promStr = md.promotion ? '成' : '',
+      role = pos.board.getRole(md.from);
     if (!role) return undefined;
     if (pos.rules === 'chushogi') {
-      if (defined(move.midStep)) {
-        const isIgui = move.to === move.from && pos.board.has(move.midStep),
-          isJitto = move.to === move.from && !isIgui,
-          midDestStr = sameDest ? sameSquareSymbol : ms(move.midStep),
-          move1 = '一歩目 ' + midDestStr + roleToFullKanji(role) + ' （←' + ms(move.from) + '）',
+      if (defined(md.midStep)) {
+        const isIgui = md.to === md.from && pos.board.has(md.midStep),
+          isJitto = md.to === md.from && !isIgui,
+          midDestStr = sameDest ? sameSquareSymbol : ms(md.midStep),
+          move1 = '一歩目 ' + midDestStr + roleToFullKanji(role) + ' （←' + ms(md.from) + '）',
           move2 =
             '二歩目 ' +
             moveDestStr +
             roleToFullKanji(role) +
             (isIgui ? '（居食い）' : isJitto ? '(じっと)' : '') +
             ' （←' +
-            ms(move.midStep) +
+            ms(md.midStep) +
             '）';
 
         return `${move1}\n${move2}`;
       }
-      return moveDestStr + roleToFullKanji(role) + promStr + ' （←' + ms(move.from) + '）';
-    } else return moveDestStr + roleToKanji(role) + promStr + '(' + makeNumberSquare(move.from) + ')';
+      return moveDestStr + roleToFullKanji(role) + promStr + ' （←' + ms(md.from) + '）';
+    } else return moveDestStr + roleToKanji(role) + promStr + '(' + makeNumberSquare(md.from) + ')';
   }
 }