+++ src/notation/kitaoKawasaki.ts
@@ -1,15 +1,15 @@
-import { Move, Square, isDrop } from '../types.js';
+import { MoveOrDrop, Square, isDrop } from '../types.js';
 import { defined } from '../util.js';
 import { Position } from '../variant/position.js';
 import { pieceCanPromote } from '../variant/util.js';
 import { aimingAt, makeNumberSquare, roleKanjiDuplicates, roleToKanji } from './util.js';
 
 // 歩-76
-export function makeKitaoKawasakiMove(pos: Position, move: Move, lastDest?: Square): string | undefined {
-  if (isDrop(move)) {
-    return roleToKanji(move.role) + '*' + makeNumberSquare(move.to);
+export function makeKitaoKawasakiMoveOrDrop(pos: Position, md: MoveOrDrop, lastDest?: Square): string | undefined {
+  if (isDrop(md)) {
+    return roleToKanji(md.role) + '*' + makeNumberSquare(md.to);
   } else {
-    const piece = pos.board.get(move.from);
+    const piece = pos.board.get(md.from);
     if (piece) {
       const roleStr = roleToKanji(piece.role).replace('成', '+'),
         ambStr = aimingAt(
@@ -17,26 +17,26 @@ export function makeKitaoKawasakiMove(pos: Position, move: Move, lastDest?: Squa
           pos.board
             .roles(piece.role, ...roleKanjiDuplicates(pos.rules)(piece.role))
             .intersect(pos.board.color(piece.color)),
-          move.to
+          md.to
         )
-          .without(move.from)
+          .without(md.from)
           .isEmpty()
           ? ''
-          : `(${makeNumberSquare(move.from)})`,
-        capture = pos.board.get(move.to),
+          : `(${makeNumberSquare(md.from)})`,
+        capture = pos.board.get(md.to),
         actionStr = !!capture ? 'x' : '-';
-      if (defined(move.midStep)) {
-        const midCapture = pos.board.get(move.midStep),
-          igui = !!midCapture && move.to === move.from;
-        if (igui) return `${roleStr}${ambStr}x!${makeNumberSquare(move.midStep)}`;
-        else if (move.to === move.from) return `--`;
+      if (defined(md.midStep)) {
+        const midCapture = pos.board.get(md.midStep),
+          igui = !!midCapture && md.to === md.from;
+        if (igui) return `${roleStr}${ambStr}x!${makeNumberSquare(md.midStep)}`;
+        else if (md.to === md.from) return `--`;
         else
           return `${roleStr}${ambStr}${!!midCapture ? 'x' : '-'}${makeNumberSquare(
-            move.midStep
-          )}${actionStr}${makeNumberSquare(move.to)}`;
+            md.midStep
+          )}${actionStr}${makeNumberSquare(md.to)}`;
       } else {
-        const destStr = (lastDest ?? pos.lastMove?.to) === move.to ? '' : makeNumberSquare(move.to),
-          promStr = move.promotion ? '+' : pieceCanPromote(pos.rules)(piece, move.from, move.to, capture) ? '=' : '';
+        const destStr = (lastDest ?? pos.lastMoveOrDrop?.to) === md.to ? '' : makeNumberSquare(md.to),
+          promStr = md.promotion ? '+' : pieceCanPromote(pos.rules)(piece, md.from, md.to, capture) ? '=' : '';
         return `${roleStr}${ambStr}${actionStr}${destStr}${promStr}`;
       }
     } else return undefined;