+++ src/notation/westernEngine.ts
@@ -1,34 +1,30 @@
-import { Move, isDrop } from '../types.js';
+import { MoveOrDrop, isDrop } from '../types.js';
 import { defined, makeSquareName } from '../util.js';
 import { Position } from '../variant/position.js';
 import { pieceCanPromote } from '../variant/util.js';
 import { aimingAt, roleToWestern } from './util.js';
 
 // P-7f
-export function makeWesternEngineMove(pos: Position, move: Move): string | undefined {
-  if (isDrop(move)) {
-    return roleToWestern(pos.rules)(move.role) + '*' + makeSquareName(move.to);
+export function makeWesternEngineMoveOrDrop(pos: Position, md: MoveOrDrop): string | undefined {
+  if (isDrop(md)) {
+    return roleToWestern(pos.rules)(md.role) + '*' + makeSquareName(md.to);
   } else {
-    const piece = pos.board.get(move.from);
+    const piece = pos.board.get(md.from);
     if (piece) {
       const roleStr = roleToWestern(pos.rules)(piece.role),
-        disambStr = aimingAt(pos, pos.board.pieces(piece.color, piece.role), move.to).without(move.from).isEmpty()
+        disambStr = aimingAt(pos, pos.board.pieces(piece.color, piece.role), md.to).without(md.from).isEmpty()
           ? ''
-          : makeSquareName(move.from),
-        toCapture = pos.board.get(move.to),
-        toStr = `${!!toCapture ? 'x' : '-'}${makeSquareName(move.to)}`;
-      if (defined(move.midStep)) {
-        const midCapture = pos.board.get(move.midStep),
-          igui = !!midCapture && move.to === move.from;
-        if (igui) return `${roleStr}${disambStr}x!${makeSquareName(move.midStep)}`;
-        else if (move.to === move.from) return `--`;
-        else return `${roleStr}${disambStr}${!!midCapture ? 'x' : '-'}${makeSquareName(move.midStep)}${toStr}`;
+          : makeSquareName(md.from),
+        toCapture = pos.board.get(md.to),
+        toStr = `${!!toCapture ? 'x' : '-'}${makeSquareName(md.to)}`;
+      if (defined(md.midStep)) {
+        const midCapture = pos.board.get(md.midStep),
+          igui = !!midCapture && md.to === md.from;
+        if (igui) return `${roleStr}${disambStr}x!${makeSquareName(md.midStep)}`;
+        else if (md.to === md.from) return `--`;
+        else return `${roleStr}${disambStr}${!!midCapture ? 'x' : '-'}${makeSquareName(md.midStep)}${toStr}`;
       } else {
-        const promStr = move.promotion
-          ? '+'
-          : pieceCanPromote(pos.rules)(piece, move.from, move.to, toCapture)
-            ? '='
-            : '';
+        const promStr = md.promotion ? '+' : pieceCanPromote(pos.rules)(piece, md.from, md.to, toCapture) ? '=' : '';
         return `${roleStr}${disambStr}${toStr}${promStr}`;
       }
     } else return undefined;