+++ test/sfen.test.ts
@@ -65,7 +65,7 @@ test('chushogi sfen', () => {
     'chushogi',
     'lfcsgekgscfl/a1b1txot1b1a/mvrhdqndhrvm/pppppppppppp/3i4i3/12/12/3I4I3/PPPPPPPPPPPP/MVRHDNQDHRVM/A1B1TOXT1B1A/LFCSGKEGSCFL b 5e 1'
   ).unwrap();
-  expect(pos2.lastMove).toEqual({ to: parseSquareName('5e') });
+  expect(pos2.lastMoveOrDrop).toEqual({ to: parseSquareName('5e') });
   expect(pos2.lastLionCapture).toEqual(parseSquareName('5e'));
 
   const pos3 = parseSfen(